# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Municipal Complaint Management API
  version: 0.1.0
paths:
...
  /notifications/me:
    get:
      tags:
      - Notifications
      summary: List my notifications
      description: Return the current user's notifications, newest first.
      operationId: list_my_notifications_notifications_me_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: unreadOnly
        in: query
        required: false
        schema:
          type: boolean
          description: Return only unread notifications
          default: false
          title: Unreadonly
        description: Return only unread notifications
      - name: type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by notification type
          title: Type
        description: Filter by notification type
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationResponse'
                title: Response List My Notifications Notifications Me Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /notifications/me/unread-count:
    get:
      tags:
      - Notifications
      summary: Count my unread notifications
      description: Cheap badge count for the navigation bar.
      operationId: get_unread_count_notifications_me_unread_count_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnreadCountResponse'
      security:
      - OAuth2PasswordBearer: []
  /notifications/{notificationId}/read:
    patch:
      tags:
      - Notifications
      summary: Mark a notification read or unread
      description: Persist read state for one notification the caller owns.
      operationId: mark_notification_read_notifications__notificationId__read_patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: notificationId
        in: path
        required: true
        schema:
          type: string
          title: Notificationid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationReadUpdate'
              default:
                isRead: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /notifications/me/read-all:
    post:
      tags:
      - Notifications
      summary: Mark all my notifications read
      operationId: mark_all_notifications_read_notifications_me_read_all_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarkAllReadResponse'
      security:
      - OAuth2PasswordBearer: []
  /notifications/{notificationId}:
    delete:
      tags:
      - Notifications
      summary: Delete a notification
      description: Remove a notification from the caller's own inbox.
      operationId: delete_notification_notifications__notificationId__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: notificationId
        in: path
        required: true
        schema:
          type: string
          title: Notificationid
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MarkAllReadResponse:
      properties:
        message:
          type: string
          title: Message
        updated:
          type: integer
          title: Updated
      type: object
      required:
      - message
      - updated
      title: MarkAllReadResponse
    NotificationReadUpdate:
      properties:
        isRead:
          type: boolean
          title: Isread
          default: true
      type: object
      title: NotificationReadUpdate
      description: Body for marking a single notification read or unread.
    NotificationResponse:
      properties:
        notificationId:
          type: string
          title: Notificationid
        message:
          type: string
          title: Message
        type:
          type: string
          title: Type
        sentAt:
          type: string
          format: date-time
          title: Sentat
        isRead:
          type: boolean
          title: Isread
        complaintId:
          type: string
          title: Complaintid
        recipientId:
          anyOf:
          - type: string
          - type: 'null'
          title: Recipientid
        readAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Readat
        priority:
          type: string
          title: Priority
          default: NORMAL
      type: object
      required:
      - notificationId
      - message
      - type
      - sentAt
      - isRead
      - complaintId
      title: NotificationResponse
...
    UnreadCountResponse:
      properties:
        unread:
          type: integer
          title: Unread
      type: object
      required:
      - unread
      title: UnreadCountResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: auth/login